+++ src/components/main/main.jsx
@@ -29,8 +29,7 @@ const Main = (props) => {
           <div className="cities__status-wrapper tabs__content">
             <b className="cities__status">No places to stay available</b>
             <p className="cities__status-description">
-              We could not find any property availbale at the moment in{` `}
-              {props.city.name}
+              We could not find any property availbale at the moment in {props.city.name}
             </p>
           </div>
         </section>
